refactor: code splitting for cart promocode components
