+++ src/components/Cart/CartCheckout/CartCheckout.scss
@@ -55,6 +55,7 @@
   &:disabled {
     background-color: $light-grey;
     color: $light-text;
+    cursor: auto;
   }
 }
