+++ src/components/Cart/CartCheckout/components/CartCheckoutPromo/handlers/onInputPromocode.ts
@@ -1,18 +1,21 @@
 /* eslint-disable no-param-reassign */
 import { PROMOCODES_DISCOUNT, PROMOCODES_NAMES } from '../../../../../../const/promocodes';
+import { resetCouponInputs } from './resetInputsHandler';
 
-export const onInputPromocode = (input: HTMLInputElement, title: HTMLElement, button: HTMLElement) => {
-  const valueInput: string = input.value.toLocaleLowerCase().trim();
+export const onInputPromocode = () => {
+  const promoDataInput = document.getElementById('coupon-input') as HTMLInputElement;
+  const valueInput: string = promoDataInput.value.toLocaleLowerCase().trim();
+  const couponBadge = document.getElementById('coupon-badge') as HTMLElement;
+  const couponBtn = document.getElementById('coupon-btn') as HTMLElement;
 
   Object.entries(PROMOCODES_NAMES).forEach(([key, value]) => {
     if (valueInput === value.toLocaleLowerCase()) {
-      title.innerHTML = `${key}  –${PROMOCODES_DISCOUNT[value]}%`;
-      button.removeAttribute('disabled');
+      couponBadge.innerHTML = `${key}  –${PROMOCODES_DISCOUNT[value]}%`;
+      couponBtn.removeAttribute('disabled');
       return;
     }
     if (valueInput === '') {
-      title.innerHTML = ``;
-      button.setAttribute('disabled', 'true');
+      resetCouponInputs();
     }
   });
 };
